requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
cssselect==1.2.0
ebaysdk==2.2.0
et-xmlfile==1.1.0
fake-useragent==1.1.3
idna==3.4
importlib-metadata==6.8.0
lxml==4.9.3
numpy==1.25.2
openpyxl==3.1.2
pandas==2.0.3
parse==1.19.1
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-html==0.10.0
six==1.16.0
soupsieve==2.4.1
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.16
w3lib==2.1.1
websockets==10.4
zipp==3.16.2
